feat: Enabled admin-only-editable paths.
This commit adds admin-only-editable paths under the working name of
x paths because we couldn't come up with a better name. All the paths
inside of the admin only paths are editable only by admins, and this is not
a database level permissions but instead using the config file so it can be
changed for all users instantaneously.